Fiddling with killing puma some more

+++ codedeploy/application_stop.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
 # Kill puma:
-kill -s 9 `ps -ef | grep puma | grep -v grep | head -n 1 | awk '{print $2}'` || echo 'puma not running -- continuing onward'
+if [ `ps -ef | grep puma | grep -v grep` ]; then
+  kill -s 9 `ps -ef | grep puma | grep -v grep | head -n 1 | awk '{print $2}'` || echo 'puma not running -- continuing onward'
+fi
